Added middleware to extend a token if necessary.
